Add several static methods to Inverse3x3Matrix
Added the transposeMatrix method from a previous project. Used this to transpose the cofacor matrix. We are now correctly creating the cofactor matrix. Renamed some variables since we are directly creating the matrix of cofactors and skipping the creation of the matrix of minors. Variables names changed to reflect this.